Added List command and Priority. Might change priority to work while CPU is running, at least will give it a try :).

+++ distrib/os/shell.js
@@ -25,7 +25,7 @@ var TSOS;
             sc = new TSOS.ShellCommand(this.shellVer, "ver", "- Displays the current version data.");
             this.commandList[this.commandList.length] = sc;
             //load
-            sc = new TSOS.ShellCommand(this.shellLoad, "load", "- Loads user input.");
+            sc = new TSOS.ShellCommand(this.shellLoad, "load", "- Loads user input. Gives Default Priority if non given");
             this.commandList[this.commandList.length] = sc;
             //clearMem
             sc = new TSOS.ShellCommand(this.shellClearMem, "clearmem", "- Kills all programs, then clears memory.");
@@ -65,6 +65,9 @@ var TSOS;
             this.commandList[this.commandList.length] = sc;
             //ShowSchedule
             sc = new TSOS.ShellCommand(this.shellShowSchedule, "currentsch", "- Show the current Scheduling Algorithm");
+            this.commandList[this.commandList.length] = sc;
+            //List
+            sc = new TSOS.ShellCommand(this.shellList, "ls", "- List all User Files in Memory(No Swap Files)");
             this.commandList[this.commandList.length] = sc;
             //Process State
             sc = new TSOS.ShellCommand(this.shellPS, "ps", "- Display all Process and their states.");
@@ -243,6 +246,7 @@ var TSOS;
             _StdOut.putText("Location: Aperture Science Laboratory");
         }
         shellLoad(args) {
+            var priority = parseInt(args, 10);
             var load;
             var val;
             var validSpace = true;
@@ -270,6 +274,14 @@ var TSOS;
                 pcb.segment = seg;
                 pcb.init();
                 pcb.location = "Memory";
+                if (Number.isInteger(priority)) {
+                    pcb.priority = priority;
+                }
+                if (isNaN(priority) || priority == null) {
+                    _StdOut.putText("Error: Priority given is Not a Number. Giving Process Default Priority");
+                    _StdOut.advanceLine();
+                    pcb.priority = defaultPriority;
+                }
                 while (start < test) {
                     let byte = val.substring(start, end);
                     _MemoryAccessor.setMAR(adr + pcb.segment.offset);
@@ -306,6 +318,14 @@ var TSOS;
                         residentlist.push(pcb);
                         TSOS.Control.addPcb(pcb);
                         _NextAvailablePID++;
+                        if (Number.isInteger(priority)) {
+                            pcb.priority = priority;
+                        }
+                        if (isNaN(priority) || priority == null) {
+                            _StdOut.putText("Error: Priority given is Not a Number. Giving Process Default Priority");
+                            _StdOut.advanceLine();
+                            pcb.priority = defaultPriority;
+                        }
                         _StdOut.putText("Valid: PID = " + pcb.pid + " Saved To Disk");
                     }
                     else {
@@ -315,6 +335,10 @@ var TSOS;
                 else {
                     _StdOut.putText("InValid");
                 }
+            }
+            //If after loading or attempting to load, a process; the scheduling algorithm is set to priority we will re-sort the resident queue
+            if (currentSchedule == "priority") {
+                _Scheduler.priority();
             }
             re.lastIndex = 0;
         }
@@ -399,6 +423,10 @@ var TSOS;
             _MemoryAccessor.clearMem();
         }
         shellQuantum(args) {
+            if (currentSchedule == "fcfs" || currentSchedule == "priority") {
+                _StdOut.putText("Cannot Change Quantum While Scheduling Algorithm is NOT Round Robin!");
+                return;
+            }
             if (args <= 0 || isNaN(args)) {
                 _StdOut.putText("Invalid Quantum!");
             }
@@ -464,6 +492,10 @@ var TSOS;
                 _StdOut.putText("Error: Missing Parameters or Disk is Not Formatted!");
             }
             else {
+                if (args[0] == "*") {
+                    _StdOut.putText("Error: Cannot Delete Swap Files!");
+                    return;
+                }
                 _krnDiskDriver.deleteFile(args);
                 _StdOut.putText("Deleted File Successfully!");
             }
@@ -472,15 +504,24 @@ var TSOS;
             if (args == "rr" || args == "fcfs" || args == "priority") {
                 currentSchedule = args;
                 if (currentSchedule == "rr") {
-                    //Do nothing, its round robin by default
+                    _StdOut.putText("Scheduling Changed to Round Robin");
+                    Quantum = 6;
+                    //Reset Quantum to Default
                 }
-                if (currentSchedule == "fcfs") {
+                else if (currentSchedule == "fcfs") {
                     //Keep it Round Robin, but with a Max Safe Integer as the Quantum
+                    _StdOut.putText("Scheduling Changed to First Come First Serve");
                     _Scheduler.fcfs();
                 }
-                if (currentSchedule == "priority" && _CPU.isExecuting) {
-                    //Still Keep it Round Robin, but re order the readyqueue by priority
+                else if (currentSchedule == "priority" && !_CPU.isExecuting) {
+                    _StdOut.putText("Scheduling Changed to Non-Preemptive Priority");
+                    _Scheduler.priority();
+                    //Still Keep it Round Robin, but re order the readyqueue by priority and make the Quantum Max Safe Integer again
+                    //Why do more work, when non-preemptive can be accomplished through a very large Quantum
                     //We only allow this when the CPU is not running, as it can cause complications re ordering the ready queue while process are running
+                }
+                else if (currentSchedule == "priority" && _CPU.isExecuting) {
+                    _StdOut.putText("Error: Cannot Switch to Non-Preemptive Priority While CPU is Running");
                 }
             }
             else {
@@ -502,6 +543,16 @@ var TSOS;
         }
         //Show all non hidden files
         shellList() {
+            let names = _krnDiskDriver.listFiles();
+            if (names.length <= 0) {
+                _StdOut.putText("Error: No Files on Disk");
+            }
+            else {
+                _StdOut.putText("Files on Disk:");
+                for (let i = 0; i < names.length; i++) {
+                    _StdOut.putText(" " + names[i] + ",");
+                }
+            }
         }
         //Copy file info to new file: first arg is name of file you want to copy, second is the file name you want to create that is a copy
         shellCopy(args) {
